requirements.txt
Flask
Flask-SocketIO
eventlet
python-binance
requests
pandas
mplfinance
numpy
# python-dotenv # Uncomment if you switch to using .env files for API keys 